Add helper functions for my pending algorithm
I'm going to need to have a sorted version for what I'm about to do. When testing, the `Array.prototype.sort()` function is actually the slowest part; which is saying something of the javascript engine!
(I don't know if it's a slow sorting algorithm or incredibly fast indexing. Now I'm a little curious…)