Add my iterator function
It's super fast! However, it requires unique values to be close by each other.
Line 70 helps us to not forget about `arr[0]`

It also the only function (except `useAssociative`) that doesn't allow a badly-written javascript engine to look through all elements between `arr[0]` and `arr[index]`, where `index` is the value we're considering is unique or not. This part about it is what inspired me to comment some code in response to your article.